+++ opcodes/microblaze-opc.h
@@ -585,8 +585,8 @@ char pvr_register_prefix[] = "rpvr";
 #define MIN_IMM6_WIDTH  ((int) 0x00000001)
 #define MAX_IMM6_WIDTH  ((int) 0x00000040)
 
-#define MIN_IMML  ((long) 0xffffff8000000000L)
-#define MAX_IMML  ((long) 0x0000007fffffffffL)
+#define MIN_IMML  ((long long) 0xffffff8000000000L)
+#define MAX_IMML  ((long long) 0x0000007fffffffffL)
 
 #endif /* MICROBLAZE_OPC */
